refactor(018/US1): split pip.rs into pip/ submodule
Split mikebom-cli/src/scan_fs/package_db/pip.rs (1965 LOC) into
five files under pip/:

  pip/mod.rs              356 LOC  orchestrator + read() + project-root
                                   walker + merge_without_override +
                                   PURL helpers + PEP 508 tokenizer
  pip/dist_info.rs        ~610 LOC Tier 1 venv site-packages walker +
                                   PipDistInfoEntry + METADATA parser +
                                   extract_license + collect_claimed_paths
  pip/poetry.rs           ~210 LOC Tier 2 poetry.lock v1/v2 parser
  pip/pipfile.rs          ~140 LOC Tier 3 Pipfile.lock parser
  pip/requirements_txt.rs ~530 LOC Tier 4 requirements*.txt parser

Net: pre-split 1965 LOC → post-split ~1846 LOC + ~250 LOC of mod
declarations / re-exports / per-module imports / per-module test
mod boilerplate. Zero production logic changed; tests preserved
verbatim and distributed per data-model.md.

Visibility ladder per spec contract (module-boundaries.md):
  - pub fn read, pub fn collect_claimed_paths re-exported from
    pip/mod.rs unchanged
  - pub(super) fn for cross-sibling parser entry points
  - private fn for parser-internal helpers
  - tests inside #[cfg(test)] mod tests at the same level as the
    code they exercise (per FR-007)

Verified:
  - cargo +stable check --workspace --tests passes clean
  - 52 pip tests + dependents all pass (cargo test pip:: 52 passed,
    0 failed)
  - All 27 byte-identity goldens (cdx + spdx-2.3 + spdx-3) unchanged:
    `MIKEBOM_UPDATE_*_GOLDENS=1` regen produces zero `git diff`
  - `cargo +stable test --workspace -- --test-threads=1` serializes
    cleanly with all targets `ok. N passed; 0 failed`

Known pre-existing flakiness: the dual_format_perf test (perf
threshold ≥ 25% reduction) is sensitive to concurrent test load.
Sometimes fails under cargo's default parallel test runner when
the system is busy; passes with `--test-threads=1` and in
isolation. This was pre-existing before the split — the baseline
run happened to pass under lighter concurrent load. Not a
regression caused by this milestone; will follow up with a
separate-PR perf-test stability fix if CI exposes the same
flakiness pattern.

US2 (npm split) and US3 (binary split) follow.
